requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
Pillow
google-generativeai
google-cloud-aiplatform
vertexai
